% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_bar}
\alias{plot_bar}
\title{Make a bar plot: measure by year (facetted using metric & category)}
\usage{
plot_bar(df, plot_title = "", measure = "value")
}
\arguments{
\item{df}{data frame with summary results}

\item{plot_title}{caption to show in plot}

\item{measure}{variable to be plotted on the y axis}
}
\description{
This is used as a for plot_value() and plot_pct()
}
\seealso{
Other functions to run dashboard visualization: \code{\link{plot_pct}},
  \code{\link{plot_segment}}, \code{\link{plot_value}},
  \code{\link{run_visual}}
}
\concept{functions to run dashboard visualization}
